comment

+++ app/Http/Controllers/Web/AlbumController.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace App\Http\Controllers\Web;
 
 use App\Models\User;
@@ -25,24 +26,22 @@ class AlbumController extends Controller
     public function create()
     {
         return view("Web.CreateAlbum");
-
     }
 
     public function show()
     {
-        $results=[];
-        $albums=Album::where('user_id',Auth::user()->id)->get();
+        $results = [];
+        $albums = Album::where('user_id', Auth::user()->id)->get();
         foreach ($albums as $album) {
-            array_push($results,[
-            'id'=>$album->id,
-             'name'=>$album->name,
-             'picturesCount'=>$album->Picture->count(),
+            array_push($results, [
+                'id' => $album->id,
+                'name' => $album->name,
+                'picturesCount' => $album->Picture->count(),
             ]);
         }
 
         // dd($results);
-        return view("dashboard",compact('results'));
-
+        return view("dashboard", compact('results'));
     }
 
     /**
@@ -50,39 +49,37 @@ class AlbumController extends Controller
      */
     public function store(StoreAlbumRequest $request)
     {
-     
-           
+
+
         $users = User::find(Auth::user()->id);
         $album = $users->Album()->create([
             'name' => $request->name,
-            ]);
+        ]);
 
-        return redirect()->route('album.show',['albumId'=>$album->id]);
-
+        return redirect()->route('album.show', ['albumId' => $album->id]);
     }
 
- 
+
 
     /**
      * Show the form for editing the specified resource.
      */
     public function edit($albumId)
     {
-        $Album=Album::where('id',$albumId)->where('user_id',Auth::user()->id)->first();
-        return view('Web.EditAlbum',['albumId'=>$albumId],compact('Album'));
+        $Album = Album::where('id', $albumId)->where('user_id', Auth::user()->id)->first();
+        return view('Web.EditAlbum', ['albumId' => $albumId], compact('Album'));
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdateAlbumRequest $request,$albumId)
+    public function update(UpdateAlbumRequest $request, $albumId)
     {
 
-        Album::where('id',$albumId)->where('user_id',Auth::user()->id)->Update([
+        Album::where('id', $albumId)->where('user_id', Auth::user()->id)->Update([
             'name' => $request->name
         ]);
         return back()->with('success', 'Edit Album successfully.');
-
     }
 
     /**
@@ -90,11 +87,12 @@ class AlbumController extends Controller
      */
     public function destroy($albumId)
     {
-        $results=Album::find($albumId)->Picture;
-        foreach ($results as $result) {
-            CustomHelpers::deletePicture($result->name);     
-           }
-        Album::where('id',$albumId)->where('user_id',Auth::user()->id)->delete();
-        return redirect()->route('albums');
+        $results = Album::find($albumId)->Picture()->count();
+        if ($results == 0) {
+            Album::where('id', $albumId)->where('user_id', Auth::user()->id)->delete();
+            return redirect()->route('albums');
+        }else{
+            return back()->with('error', 'have pictures.');
+        }
     }
 }
